added generate exception to parameters
